Split EBS Volume & EBS Snapshot functions
