+++ aws-inventory/cloudformation/Inventory-Template.yaml
@@ -1176,12 +1176,12 @@ Resources:
           ERROR_QUEUE: !Ref ErrorHandlerEventQueue
       # Tags inherited from Stack
 
-  EBSInventoryLambdaFunction:
+  EBSVolInventoryLambdaFunction:
     Type: AWS::Lambda::Function
     Properties:
-      FunctionName: !Sub "${AWS::StackName}-ebs-inventory"
+      FunctionName: !Sub "${AWS::StackName}-ebs-vol-inventory"
       Description: Inventory EBS Volumes & Snapshots
-      Handler: inventory-ebs.lambda_handler
+      Handler: inventory-ebs-volume.lambda_handler
       Runtime: python3.6
       Layers:
         - !Ref AWSInventoryLambdaLayer
@@ -1201,6 +1201,30 @@ Resources:
           ERROR_QUEUE: !Ref ErrorHandlerEventQueue
       # Tags inherited from Stack
 
+  EBSSnapInventoryLambdaFunction:
+    Type: AWS::Lambda::Function
+    Properties:
+      FunctionName: !Sub "${AWS::StackName}-ebs-snap-inventory"
+      Description: Inventory EBS Volumes & Snapshots
+      Handler: inventory-ebs-snapshot.lambda_handler
+      Runtime: python3.6
+      Layers:
+        - !Ref AWSInventoryLambdaLayer
+      Timeout: 300
+      MemorySize: !Ref pLargeLambdaSize
+      Role: !GetAtt InventoryLambdaRole.Arn
+      Code:
+        S3Bucket: !Ref pBucketName
+        S3Key: !Sub ${pLambdaZipFile}
+      Environment:
+        Variables:
+          ROLE_SESSION_NAME: !Ref AWS::StackName
+          INVENTORY_BUCKET: !Ref pBucketName
+          ACCOUNT_TABLE: !Ref AccountDBTable
+          VPC_TABLE: !Ref VpcInventoryDBTable
+          ROLE_NAME: !Ref pRoleName
+          ERROR_QUEUE: !Ref ErrorHandlerEventQueue
+      # Tags inherited from Stack
 
   ELBInventoryLambdaFunction:
     Type: AWS::Lambda::Function
@@ -1817,18 +1841,33 @@ Resources:
       Protocol: lambda
       TopicArn: !Ref 'TriggerAccountInventoryFunctionTopic'
 
-  EBSInventoryLambdaFunctionPermission:
+  EBSVolInventoryLambdaFunctionPermission:
     Type: AWS::Lambda::Permission
     Properties:
-      FunctionName: !GetAtt EBSInventoryLambdaFunction.Arn
+      FunctionName: !GetAtt EBSVolInventoryLambdaFunction.Arn
       Principal: sns.amazonaws.com
       SourceArn: !Ref TriggerAccountInventoryFunctionTopic
       Action: lambda:invokeFunction
 
-  EBSInventoryTopicToLambdaSubscription:
+  EBSVolInventoryTopicToLambdaSubscription:
     Type: AWS::SNS::Subscription
     Properties:
-      Endpoint: !GetAtt [EBSInventoryLambdaFunction, Arn]
+      Endpoint: !GetAtt [EBSVolInventoryLambdaFunction, Arn]
+      Protocol: lambda
+      TopicArn: !Ref 'TriggerAccountInventoryFunctionTopic'
+
+  EBSSnapInventoryLambdaFunctionPermission:
+    Type: AWS::Lambda::Permission
+    Properties:
+      FunctionName: !GetAtt EBSSnapInventoryLambdaFunction.Arn
+      Principal: sns.amazonaws.com
+      SourceArn: !Ref TriggerAccountInventoryFunctionTopic
+      Action: lambda:invokeFunction
+
+  EBSSnapInventoryTopicToLambdaSubscription:
+    Type: AWS::SNS::Subscription
+    Properties:
+      Endpoint: !GetAtt [EBSSnapInventoryLambdaFunction, Arn]
       Protocol: lambda
       TopicArn: !Ref 'TriggerAccountInventoryFunctionTopic'
 
@@ -2075,7 +2114,8 @@ Resources:
                   [ "...", "${AWS::StackName}-cloudfront-inventory", { "stat": "Sum", "period": 604800, "label": "cloudfront-inventory" } ],
                   [ "...", "${AWS::StackName}-cloudtrail-inventory", { "stat": "Sum", "period": 604800, "label": "cloudtrail-inventory" } ],
                   [ "...", "${AWS::StackName}-dx-inventory", { "stat": "Sum", "period": 604800, "label": "dx-inventory" } ],
-                  [ "...", "${AWS::StackName}-ebs-inventory", { "stat": "Sum", "period": 604800, "label": "ebs-inventory" } ],
+                  [ "...", "${AWS::StackName}-ebs-vol-inventory", { "stat": "Sum", "period": 604800, "label": "ebs-vol-inventory" } ],
+                  [ "...", "${AWS::StackName}-ebs-snap-inventory", { "stat": "Sum", "period": 604800, "label": "ebs-snap-inventory" } ],
                   [ "...", "${AWS::StackName}-ecr-inventory", { "stat": "Sum", "period": 604800, "label": "ecr-inventory" } ],
                   [ "...", "${AWS::StackName}-ecs-inventory", { "stat": "Sum", "period": 604800, "label": "ecs-inventory" } ],
                   [ "...", "${AWS::StackName}-elb-inventory", { "stat": "Sum", "period": 604800, "label": "elb-inventory" } ],
@@ -2127,7 +2167,8 @@ Resources:
                   [ "...", "${AWS::StackName}-cloudfront-inventory", { "stat": "Sum", "period": 604800, "label": "cloudfront-inventory" } ],
                   [ "...", "${AWS::StackName}-cloudtrail-inventory", { "stat": "Sum", "period": 604800, "label": "cloudtrail-inventory" } ],
                   [ "...", "${AWS::StackName}-dx-inventory", { "stat": "Sum", "period": 604800, "label": "dx-inventory" } ],
-                  [ "...", "${AWS::StackName}-ebs-inventory", { "stat": "Sum", "period": 604800, "label": "ebs-inventory" } ],
+                  [ "...", "${AWS::StackName}-ebs-vol-inventory", { "stat": "Sum", "period": 604800, "label": "ebs-vol-inventory" } ],
+                  [ "...", "${AWS::StackName}-ebs-snap-inventory", { "stat": "Sum", "period": 604800, "label": "ebs-snap-inventory" } ],
                   [ "...", "${AWS::StackName}-ecr-inventory", { "stat": "Sum", "period": 604800, "label": "ecr-inventory" } ],
                   [ "...", "${AWS::StackName}-ecs-inventory", { "stat": "Sum", "period": 604800, "label": "ecs-inventory" } ],
                   [ "...", "${AWS::StackName}-elb-inventory", { "stat": "Sum", "period": 604800, "label": "elb-inventory" } ],
@@ -2179,7 +2220,8 @@ Resources:
                   [ "...", "${AWS::StackName}-cloudfront-inventory", { "stat": "Sum", "period": 900, "label": "cloudfront-inventory" } ],
                   [ "...", "${AWS::StackName}-cloudtrail-inventory", { "stat": "Sum", "period": 900, "label": "cloudtrail-inventory" } ],
                   [ "...", "${AWS::StackName}-dx-inventory", { "stat": "Sum", "period": 900, "label": "dx-inventory" } ],
-                  [ "...", "${AWS::StackName}-ebs-inventory", { "stat": "Sum", "period": 900, "label": "ebs-inventory" } ],
+                  [ "...", "${AWS::StackName}-ebs-vol-inventory", { "stat": "Sum", "period": 900, "label": "ebs-vol-inventory" } ],
+                  [ "...", "${AWS::StackName}-ebs-snap-inventory", { "stat": "Sum", "period": 900, "label": "ebs-snap-inventory" } ],
                   [ "...", "${AWS::StackName}-ecr-inventory", { "stat": "Sum", "period": 900, "label": "ecr-inventory" } ],
                   [ "...", "${AWS::StackName}-ecs-inventory", { "stat": "Sum", "period": 900, "label": "ecs-inventory" } ],
                   [ "...", "${AWS::StackName}-elb-inventory", { "stat": "Sum", "period": 900, "label": "elb-inventory" } ],
