+++ aws-inventory/lambda/inventory-ebs-volume.py
@@ -20,9 +20,7 @@ logging.getLogger('urllib3').setLevel(logging.WARNING)
 
 
 VOLUME_RESOURCE_PATH = "ec2/volume"
-SNAPSHOT_RESOURCE_PATH = "ec2/snapshot"
 VOLUME_TYPE = "AWS::EC2::Volume"
-SNAPSHOT_TYPE = "AWS::EC2::Snapshot"
 
 
 def lambda_handler(event, context):
@@ -34,12 +32,14 @@ def lambda_handler(event, context):
         target_account = AWSAccount(message['account_id'])
         for r in target_account.get_regions():
             discover_volumes(target_account, r)
-            discover_snapshots(target_account, r)
 
     except AntiopeAssumeRoleError as e:
         logger.error("Unable to assume role into account {}({})".format(target_account.account_name, target_account.account_id))
         return()
     except ClientError as e:
+        if e.response['Error']['Code'] == 'UnauthorizedOperation':
+            logger.error("Antiope doesn't have proper permissions to this account")
+            return(event)
         logger.critical("AWS Error getting info for {}: {}".format(message['account_id'], e))
         capture_error(message, context, e, "ClientError for {}: {}".format(message['account_id'], e))
         raise
@@ -82,37 +82,7 @@ def discover_volumes(account, region):
         save_resource_to_s3(VOLUME_RESOURCE_PATH, resource_item['resourceId'], resource_item)
 
 
-def discover_snapshots(account, region):
-    '''
-        Discover EBS Snapshots owned by this account
 
-    '''
-
-    snapshots = []
-
-    ec2_client = account.get_client('ec2', region=region)
-    response = ec2_client.describe_snapshots( OwnerIds=[account.account_id ])
-    while 'NextToken' in response:  # Gotta Catch 'em all!
-        snapshots += response['Snapshots']
-        response = ec2_client.describe_snapshots(OwnerIds=[account.account_id ], NextToken=response['NextToken'])
-    snapshots += response['Snapshots']
-
-    for snap in snapshots:
-        resource_item = {}
-        resource_item['awsAccountId']                   = account.account_id
-        resource_item['awsAccountName']                 = account.account_name
-        resource_item['resourceType']                   = SNAPSHOT_TYPE
-        resource_item['source']                         = "Antiope"
-        resource_item['awsRegion']                      = region
-        resource_item['configurationItemCaptureTime']   = str(datetime.datetime.now())
-        resource_item['configuration']                  = snap
-        if 'Tags' in snap:
-            resource_item['tags']                       = parse_tags(snap['Tags'])
-        resource_item['supplementaryConfiguration']     = {}
-        resource_item['resourceId']                     = snap['SnapshotId']
-        resource_item['resourceName']                   = snap['SnapshotId']
-        resource_item['errors']                         = {}
-        save_resource_to_s3(SNAPSHOT_RESOURCE_PATH, resource_item['resourceId'], resource_item)
 
 
 def json_serial(obj):
